refactored the main page delete button
