+++ ui/app/globals.css
@@ -48,9 +48,9 @@
   @apply relative inline-flex items-center justify-center px-8 py-4 text-lg font-bold 
          rounded-lg shadow-lg border transition-all duration-300 ease-in-out;
 
-  background-color: #3c1642; /* base purple */
-  border: 2px solid #4a2f6f; /* darker purple border */
-  text-shadow: 0 1px 0 rgba(0,0,0,0.3); /* subtle depth */
+  background-color: #3c1642;
+  border: 2px solid #4a2f6f;
+  text-shadow: 0 1px 0 rgba(0,0,0,0.3); 
   user-select: none;
   background-size: cover;
   background-repeat: no-repeat;
@@ -62,7 +62,7 @@
 .btn-primary:hover {
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.25), 0 0 8px rgba(75, 47, 111, 0.5); 
   transform: translateY(-1px);
-  border-color: #3e2660; /* slightly darker border on hover */
+  border-color: #3e2660; 
 }
 
 .btn-primary:active {
@@ -76,6 +76,20 @@
   outline-offset: 2px;
 }
 
+
+.btn-warning {
+  @apply relative inline-flex items-center justify-center px-8 py-4 text-lg font-bold 
+         rounded-lg shadow-lg border transition-all duration-300 ease-in-out;
+
+  background-color: #a80d18; 
+  colour: #eff1ed;
+  border: 2px solid #8f0a13; 
+  user-select: none;
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-position: center;
+  select: none;
+}
